+++ envoy/http/async_client.h
@@ -46,80 +46,6 @@ public:
   virtual void onBelowWriteBufferLowWatermark() PURE;
 };
 
-class DecoderFilterSidestreamWatermarkCallbacks : public Http::SidestreamWatermarkCallbacks {
-public:
-  DecoderFilterSidestreamWatermarkCallbacks() = default;
-
-  void setDecoderFilterCallbacks(Http::StreamDecoderFilterCallbacks* callback) {
-    callback_ = callback;
-  }
-  void onAboveWriteBufferHighWatermark() final {
-    if (callback_ != nullptr) {
-      // TODO(tyxia) Change to Envoy bug
-      callback_->onDecoderFilterAboveWriteBufferHighWatermark();
-    } else {
-    }
-  }
-
-  void onBelowWriteBufferLowWatermark() final {
-    if (callback_ != nullptr) {
-      callback_->onDecoderFilterBelowWriteBufferLowWatermark();
-    } else {
-      // TODO(tyxia) Change to Envoy bug
-    }
-  }
-
-  void resetFilterCallbacks() { callback_ = nullptr; }
-
-private:
-  Http::StreamDecoderFilterCallbacks* callback_ = nullptr;
-};
-
-class StreamFilterSidestreamWatermarkCallbacks : public Http::SidestreamWatermarkCallbacks {
-public:
-  StreamFilterSidestreamWatermarkCallbacks() = default;
-
-  void onAboveWriteBufferHighWatermark() final {
-    if (decode_callback_ != nullptr) {
-      decode_callback_->onDecoderFilterAboveWriteBufferHighWatermark();
-    } else {
-    }
-
-    if (encode_callback_ != nullptr) {
-      encode_callback_->onEncoderFilterAboveWriteBufferHighWatermark();
-    } else {
-    }
-  }
-
-  void onBelowWriteBufferLowWatermark() final {
-    if (decode_callback_ != nullptr) {
-      decode_callback_->onDecoderFilterBelowWriteBufferLowWatermark();
-    }
-
-    if (encode_callback_ != nullptr) {
-      encode_callback_->onEncoderFilterBelowWriteBufferLowWatermark();
-    } else {
-    }
-  }
-
-  void setDecoderFilterCallbacks(Http::StreamDecoderFilterCallbacks* decode_callback) {
-    decode_callback_ = decode_callback;
-  }
-
-  void setEncoderFilterCallbacks(Http::StreamEncoderFilterCallbacks* encode_callback) {
-    encode_callback_ = encode_callback;
-  }
-
-  void resetFilterCallbacks() {
-    decode_callback_ = nullptr;
-    encode_callback_ = nullptr;
-  }
-
-private:
-  Http::StreamDecoderFilterCallbacks* decode_callback_ = nullptr;
-  Http::StreamEncoderFilterCallbacks* encode_callback_ = nullptr;
-};
-
 /**
  * Supports sending an HTTP request message and receiving a response asynchronously.
  */
