+++ envoy/http/filter.h
@@ -1255,6 +1255,5 @@ public:
    */
   virtual Event::Dispatcher& dispatcher() PURE;
 };
-
 } // namespace Http
 } // namespace Envoy
