+++ source/extensions/filters/http/ext_proc/client.h
@@ -9,6 +9,8 @@
 #include "envoy/service/ext_proc/v3/external_processor.pb.h"
 #include "envoy/stream_info/stream_info.h"
 
+#include "source/common/http/async_client_utility.h"
+
 namespace Envoy {
 namespace Extensions {
 namespace HttpFilters {
